+++ src/main/java/subway/line/domain/SubwayMap.java
@@ -36,41 +36,6 @@ public class SubwayMap {
     }
 
     public DirectedSections getShortestPath(Station source, Station destination) {
-        List<Section> sections = findAllSections();
-        DijkstraShortestPath<Station, DefaultWeightedEdge> graph = createGraph(sections);
-        List<Station> orderedStations = graph.getPath(source, destination).getVertexList();
-        List<DirectedSection> directedSections = new ArrayList<>();
-
-        for (int i = 0; i < orderedStations.size() - 1; i++) {
-            Station sourceStation = orderedStations.get(i);
-            Station targetStation = orderedStations.get(i + 1);
-            DirectedSection directedSection = new DirectedSection(findSection(sections, sourceStation, targetStation), sourceStation);
-            directedSections.add(directedSection);
-        }
-        return new DirectedSections(directedSections,
-                source,
-                destination,
-                graph,
-                ExtraFare.of(this, directedSections));
-    }
-
-    private DijkstraShortestPath<Station, DefaultWeightedEdge> createGraph(List<Section> sections) {
-        WeightedMultigraph<Station, DefaultWeightedEdge> graph =
-                new WeightedMultigraph<>(DefaultWeightedEdge.class);
-
-        for (Section section : sections) {
-            graph.addVertex(section.getUpStation());
-            graph.addVertex(section.getDownStation());
-            graph.setEdgeWeight(graph.addEdge(section.getUpStation(), section.getDownStation()), section.getDistance());
-        }
-
-        return new DijkstraShortestPath<>(graph);
-    }
-
-    private Section findSection(List<Section> sections, Station stationA, Station stationB) {
-        return sections.stream()
-                .filter(section -> section.contains(stationA) && section.contains(stationB))
-                .findAny()
-                .orElseThrow(IllegalStateException::new);
+        return new DirectedSections(this, source, destination);
     }
 }
